requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pdfplumber>=0.11.0
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
numpy>=1.24.0
